+++ CHANGELOG.rst
@@ -34,6 +34,7 @@ Changed
   vice-versa).
 - junc_image_req_ldst.py got some additional columns to manage the Science Team
   evaluation of images acquired from Image Requests.
+- create_pano.py's make_pano_product() function renamed to make_pano_record().
 
 Added
 ^^^^^
@@ -41,11 +42,13 @@ Added
 - Association table junc_image_pano created which provides a many-to-many
   connection between ImageRecords and PanoRecords and added bidirectional
   relationship entries to each table.
-- pano_records table now has azimuth and elevation angle min/max values to indicate
+- pano_records table now has pan and tilt angle min/max values to indicate
   angular range of panorama coverage.
 - image_requests.py - "Acquired," "Not Acquired," and "Not Planned" statuses added to
   enum.
 - ptu_records.py - Tables to record the pan and tilt of the rover's pan-tilt-unit (PTU).
+- create_pano.py - updated to correctly add PanoRecord associations, can now query
+  database for ImageRecords.
 
 Removed
 ^^^^^^^